+++ examples/thread-loader/webpack.config.js
@@ -8,7 +8,7 @@ module.exports = {
     entry: './src/index.ts',
     output: { filename: 'dist/index.js' },
     module: {
-        rules: {
+        rules: [{
             test: /\.tsx?$/,
             use: [
                 { loader: 'cache-loader' },
@@ -26,12 +26,12 @@ module.exports = {
                     }
                 }
             ]
-        }
+        }]
     },
     resolve: {
         extensions: ['.ts', '.tsx', 'js']
     },
     plugins: [
-        new ForkTsCheckerWebpackPlugin()
+        new ForkTsCheckerWebpackPlugin({ checkSyntacticErrors: true })
     ]
 };