+++ src/config.ts
@@ -24,7 +24,7 @@ export function getConfigFile(
     log: logger.Logger,
     compilerDetailsLogMessage: string
 ) {
-    const configFilePath = findConfigFile(compiler, path.dirname(loader.resourcePath), loaderOptions.configFileName);
+    const configFilePath = findConfigFile(compiler, path.dirname(loader.resourcePath), loaderOptions.configFile);
     let configFileError: WebpackError | undefined;
     let configFile: ConfigFile;
 
@@ -69,23 +69,51 @@ export function getConfigFile(
     };
 }
 
+
 /**
- * The tsconfig.json is found using the same method as `tsc`, starting in the current directory
- * and continuing up the parent directory chain.
+ * Find a tsconfig file by name or by path.
+ * By name, the tsconfig.json is found using the same method as `tsc`, starting in the current
+ * directory and continuing up the parent directory chain.
+ * By path, the file will be found by resolving the given path relative to the requesting entry file.
+ * 
+ * @param compiler The TypeScript compiler instance
+ * @param requestDirPath The directory in which the entry point requesting the tsconfig.json lies
+ * @param configFile The tsconfig file name to look for or a path to that file
+ * @return The absolute path to the tsconfig file, undefined if none was found.
  */
-function findConfigFile(compiler: typeof typescript, searchPath: string, configFileName: string): string | undefined {
-    while (true) {
-        const fileName = path.join(searchPath, configFileName);
-        if (compiler.sys.fileExists(fileName)) {
-            return fileName;
-        }
-        const parentPath = path.dirname(searchPath);
-        if (parentPath === searchPath) {
-            break;
-        }
-        searchPath = parentPath;
+function findConfigFile(compiler: typeof typescript, requestDirPath: string, configFile: string): string | undefined {
+    // If `configFile` is an absolute path, return it right away
+    if (path.isAbsolute(configFile)) {
+        return compiler.sys.fileExists(configFile)
+            ? configFile
+            : undefined;
     }
-    return undefined;
+
+    // If `configFile` is a relative path, resolve it.
+    // We define a relative path as: starts with
+    // one or two dots + a common directory delimiter
+    if (configFile.match(/^\.\.?(\/|\\)/)) {
+        const resolvedPath = path.resolve(requestDirPath, configFile);
+        return compiler.sys.fileExists(resolvedPath)
+            ? resolvedPath
+            : undefined;
+
+    // If `configFile` is a file name, find it in the directory tree
+    } else {
+        while (true) {
+            const fileName = path.join(requestDirPath, configFile);
+            if (compiler.sys.fileExists(fileName)) {
+                return fileName;
+            }
+            const parentPath = path.dirname(requestDirPath);
+            if (parentPath === requestDirPath) {
+                break;
+            }
+            requestDirPath = parentPath;
+        }
+
+        return undefined;
+    }
 }
 
 export function getConfigParseResult(